upgraded to gpui-2.2.0 tag 1.9.0 (v2)

+++ crates/gpui-ios/src/ios/display.rs
@@ -77,7 +77,7 @@ impl Drop for IosDisplay {
 
 impl PlatformDisplay for IosDisplay {
     fn id(&self) -> DisplayId {
-        DisplayId::new(self.screen as u32)
+        DisplayId::new(self.screen as u64)
     }
 
     fn uuid(&self) -> anyhow::Result<Uuid> {
